+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Store Aquos error in booking entry as _aquos_status 
+
 ### 1.7.0: July 6th, 2021
 * Refactoring of dashboard calculation, now calculated with cron event